+++ src/com/srscicomp/fc/ui/FGNSymbolCard.java
@@ -15,10 +15,7 @@ import javax.swing.text.Document;
 
 import com.srscicomp.common.g2dutil.Marker;
 import com.srscicomp.common.ui.MaxLengthDocumentFilter;
-import com.srscicomp.fc.fig.FGNPlottable;
-import com.srscicomp.fc.fig.FGNodeType;
-import com.srscicomp.fc.fig.FGraphicModel;
-import com.srscicomp.fc.fig.SymbolNode;
+import com.srscicomp.fc.fig.*;
 import com.srscicomp.fc.uibase.MeasureEditor;
 
 /**
@@ -30,20 +27,30 @@ import com.srscicomp.fc.uibase.MeasureEditor;
  * that exposes the symbol's draw styles. If the symbol node supports rendering a single character centered on the 
  * marker symbol, a short text field is included to specify that character. Font characteristics are set by the parent 
  * node's text styling attributes.</p>
+ *
+ * <p><b></b>Special use case - 3D scatter plot</b>. The {@link Scatter3DNode} specifies a minimum symbol size, while
+ * the size of its component {@link SymbolNode} is the maximum symbol size. Logically, the minimum symbol size widget
+ * should be included in this editor, but it's not a property of {@link SymbolNode}. To accommodate this special case,
+ * this editor can be configured to include a minimum symbol size widget, and the tool tip for the regular size widget
+ * is adjusted accordingly. In this configuration, the parent of the edited {@link SymbolNode} must be an instance of
+ * {@link Scatter3DNode}.</p>
  * 
  * @author sruffner
  */
 class FGNSymbolCard extends JPanel implements ActionListener, FocusListener
 {
    /** Construct the symbol properties editor such that it exposes all properties of a {@link SymbolNode}. */
-   FGNSymbolCard() { this(false, false); }
-   
+   FGNSymbolCard() { this(false, false, false); }
+
+
    /** 
     * Construct the symbol node properties editor. 
     * @param omitSymChar True to omit the short text field exposing the symbol character.
     * @param omitFillC True to omit the fill color picker from the embedded draw styles editor.
+    * @param includeMinSize If true, include minimum symbol size widget to handle the component {@link SymbolNode} for
+    * a 3D scatter plot. In this case, the edited node must be a {@link Scatter3DNode} rather than {@link SymbolNode}.
     */
-   FGNSymbolCard(boolean omitSymChar, boolean omitFillC)
+   FGNSymbolCard(boolean omitSymChar, boolean omitFillC, boolean includeMinSize)
    {
       super();
       setOpaque(false);
@@ -53,11 +60,18 @@ class FGNSymbolCard extends JPanel implements ActionListener, FocusListener
       typeCombo.addActionListener(this);
       add(typeCombo);
       
-      sizeEditor = new MeasureEditor(0, FGraphicModel.getSizeConstraints(FGNodeType.SHAPE));
-      sizeEditor.setToolTipText("Enter size of symbols's bounding box");
+      sizeEditor = new MeasureEditor(0, SymbolNode.SYMBOLSIZECONSTRAINTS);
+      sizeEditor.setToolTipText(includeMinSize ? "Maximum (or fixed) symbol size" : "Size of symbol bounding box");
       sizeEditor.addActionListener(this);
       add(sizeEditor);
-      
+
+      if(includeMinSize)
+      {
+         minSizeEditor =new MeasureEditor(0, SymbolNode.SYMBOLSIZECONSTRAINTS);
+         minSizeEditor.setToolTipText("Minimum symbol size (must be less than max size)");
+         minSizeEditor.addActionListener(this);
+         add(minSizeEditor);
+      }
       if(!omitSymChar)
       {
          charField = new JTextField(2);
@@ -81,8 +95,12 @@ class FGNSymbolCard extends JPanel implements ActionListener, FocusListener
       
       // left-right constraints. Assume widest row is the one containing the draw style editor. 
       layout.putConstraint(SpringLayout.WEST, typeCombo, 0, SpringLayout.WEST, this);
-      layout.putConstraint(SpringLayout.WEST, sizeEditor, GAP, SpringLayout.EAST, typeCombo);
-      if(charField != null) layout.putConstraint(SpringLayout.WEST, charField, GAP*3, SpringLayout.EAST, sizeEditor);
+      if(minSizeEditor != null)
+         layout.putConstraint(SpringLayout.WEST, minSizeEditor, GAP, SpringLayout.EAST, typeCombo);
+      layout.putConstraint(SpringLayout.WEST, sizeEditor, GAP, SpringLayout.EAST,
+            minSizeEditor != null ? minSizeEditor : typeCombo);
+      if(charField != null)
+         layout.putConstraint(SpringLayout.WEST, charField, GAP*3, SpringLayout.EAST, sizeEditor);
       
       layout.putConstraint(SpringLayout.WEST, drawStyleEditor, 0, SpringLayout.WEST, this);
       layout.putConstraint(SpringLayout.EAST, this, 0, SpringLayout.EAST, drawStyleEditor);
@@ -95,16 +113,29 @@ class FGNSymbolCard extends JPanel implements ActionListener, FocusListener
       
       String vc = SpringLayout.VERTICAL_CENTER;
       layout.putConstraint(vc, sizeEditor, 0, vc, typeCombo);
+      if(minSizeEditor != null) layout.putConstraint(vc, minSizeEditor, 0, vc, typeCombo);
       if(charField != null) layout.putConstraint(vc, charField, 0, vc, typeCombo);
    }
 
+   /**
+    * Reload editor IAW contents of specified symbol node.
+    *
+    * @param sn The edited symbol node. If the editor is configured with the minimum symbol size widget, then the
+    * parent of the symbol node must be an instance of {@link Scatter3DNode}.
+    */
    public void reload(SymbolNode sn)
    {
-      if(sn != null) symbol = sn;
+      if(sn != null && ((minSizeEditor == null) || (sn.getParent() instanceof Scatter3DNode)))
+         symbol = sn;
       if(symbol == null) return;
       
       typeCombo.setSelectedItem(symbol.getType());
       sizeEditor.setMeasure(symbol.getSize());
+      if(minSizeEditor != null)
+      {
+         Scatter3DNode scat3d = (Scatter3DNode) symbol.getParent();
+         minSizeEditor.setMeasure(scat3d.getMinSymbolSize());
+      }
       if(charField != null) charField.setText(symbol.getCharacter());
       
       drawStyleEditor.loadGraphicNode(symbol);
@@ -139,11 +170,12 @@ class FGNSymbolCard extends JPanel implements ActionListener, FocusListener
       if(symbol == null) return;
       Object src = e.getSource();
 
+      boolean ok = true;
       if(charField != null && src == charField)
       {
          if(!symbol.setTitle(charField.getText()))
          {
-            Toolkit.getDefaultToolkit().beep();
+            ok = false;
             charField.setText(symbol.getTitle());
          }
       }
@@ -151,18 +183,37 @@ class FGNSymbolCard extends JPanel implements ActionListener, FocusListener
       {
          if(!symbol.setType((Marker)typeCombo.getSelectedItem()))
          {
-            Toolkit.getDefaultToolkit().beep();
+            ok = false;
             typeCombo.setSelectedItem(symbol.getType());
          }
       }
       else if(src == sizeEditor)
       {
+         if(minSizeEditor != null)
+         {
+            Scatter3DNode scat3d = (Scatter3DNode) symbol.getParent();
+            if(!scat3d.setMaxSymbolSize(sizeEditor.getMeasure()))
+            {
+               ok = false;
+               sizeEditor.setMeasure(scat3d.getMaxSymbolSize());
+            }
+         }
          if(!symbol.setSize(sizeEditor.getMeasure()))
          {
-            Toolkit.getDefaultToolkit().beep();
+            ok = false;
             sizeEditor.setMeasure(symbol.getSize());
          }
       }
+      else if(minSizeEditor != null && src == minSizeEditor)
+      {
+         Scatter3DNode scat3d = (Scatter3DNode) symbol.getParent();
+         if(!scat3d.setMinSymbolSize(minSizeEditor.getMeasure()))
+         {
+            ok = false;
+            minSizeEditor.setMeasure(scat3d.getMinSymbolSize());
+         }
+      }
+      if(!ok) Toolkit.getDefaultToolkit().beep();
    }
 
    /** The current symbol node being edited. If null, editor is non-functional. */
@@ -173,6 +224,12 @@ class FGNSymbolCard extends JPanel implements ActionListener, FocusListener
 
    /** Customized component for editing the symbol marker's size. */
    private MeasureEditor sizeEditor = null;
+
+   /**
+    * Customized component for editing minimum symbol size -- applicable only when the edited node is a 3D scatter
+    * plot instead of a symbol node.
+    */
+   private MeasureEditor minSizeEditor = null;
 
    /** 
     * Text field for specifying a single character to be drawn centered inside symbol. Will be null if omitted from